on the way to arm
