+++ armExampleProject.deep
@@ -1,0 +1,36 @@
+#deep-1
+
+meta {
+	version = "2013-09-25";
+	description = "Example project file";
+}
+
+project ExampleProjectMpc555 {
+	libpath = "../ch.ntb.inf.deep.trglib", "../ch.ntb.inf.junitTarget";
+#	libpath = "C:/Program Files/Java/jre6/lib/rt.jar","../ch.ntb.inf.deep.trglib";
+#	libpath = "../ch.ntb.inf.deep.trglib", "../ch.ntb.inf.junitTarget","C:/Program Files/Java/jre6/lib/rt.jar";
+	
+	boardtype = colibri_iMX6;
+	ostype = ntbSTS_arm;
+#	programmertype = ntbMpc555UsbBdi;
+	
+#	rootclasses = "java/lang/Object";
+#	rootclasses = "test/Test1";
+	rootclasses = "ch/ntb/inf/deep/runtime/mpc555/test/BlinkerApplication","ch/ntb/inf/deep/runtime/mpc555/test/OutTest1";
+#	rootclasses = "ch/ntb/inf/deep/runtime/mpc555/demo/SystemOutDemo1";
+#	rootclasses = "ch/ntb/inf/deep/runtime/mpc555/test/OutTest1";
+#	rootclasses = "ch/ntb/inf/deep/runtime/mpc555/test/BlinkerApplication";
+#	rootclasses = "ch/ntb/inf/deep/runtime/mpc555/test/ActionableTest2";	
+#	rootclasses = "ch/ntb/inf/deep/runtime/mpc555/test/ExceptionTest2";	
+#	rootclasses = "ch.ntb.inf.deep.comp.targettest.various.VariousTest1";
+#	rootclasses = "ch.ntb.inf.deep.trglib.targettest.SystemTest";
+#	rootclasses = "ch/ntb/inf/deep/runtime/mpc555/test/HeapTest3";
+#	rootclasses = "ch/ntb/inf/deep/runtime/mpc555/demo/SystemInOutReflector";
+#	rootclasses = "ch/ntb/inf/deep/runtime/mpc555/driver/ffs/FileTransfer";
+
+#	imgfile = "out/exampleProject.bin"
+#	imgformat = binary;
+
+#	tctfile = "tct/targetCommands.dtct";
+
+}
