+++ junitHostTest/ch/ntb/inf/deep/comp/hosttest/cgPPC/cgPPC02.java
@@ -25,7 +25,7 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 
 import ch.ntb.inf.deep.cg.Code32;
-import ch.ntb.inf.deep.cg.ppc.RegAllocator;
+import ch.ntb.inf.deep.cg.ppc.RegAllocatorPPC;
 import ch.ntb.inf.deep.classItems.CFR;
 import ch.ntb.inf.deep.classItems.Class;
 import ch.ntb.inf.deep.config.Configuration;
@@ -50,7 +50,7 @@ public class cgPPC02 extends TestCgPPC {
 		assertNull("wrong join", getJoin(1));
 		assertNull("wrong join", getJoin(2));
 		assertTrue("wrong join", checkJoin(getJoin(3), 3, 8, vol, false));
-		for (int i = 4; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 4; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 7, false, false));
 	}
@@ -61,7 +61,7 @@ public class cgPPC02 extends TestCgPPC {
 		assertNull("wrong join", getJoin(0));
 		assertNull("wrong join", getJoin(1));
 		assertTrue("wrong join", checkJoin(getJoin(2), 3, 9, vol, false));
-		for (int i = 3; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 3; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 6, false, false));
 	}
@@ -73,7 +73,7 @@ public class cgPPC02 extends TestCgPPC {
 		assertNull("wrong join", getJoin(1));
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 18, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(3), 1, 19, vol, false));
-		for (int i = 4; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 4; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 18, false, false));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 19, false, false));
@@ -87,7 +87,7 @@ public class cgPPC02 extends TestCgPPC {
 		assertNull("wrong join", getJoin(2));
 		assertNull("wrong join", getJoin(3));
 		assertTrue("wrong join", checkJoin(getJoin(4), 4, 8, vol, false));
-		for (int i = 5; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 5; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 7, false, false));
 	}
@@ -100,7 +100,7 @@ public class cgPPC02 extends TestCgPPC {
 		assertNull("wrong join", getJoin(2));
 		assertTrue("wrong join", checkJoin(getJoin(3), 0, 7, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(4), 0, 7, vol, false));
-		for (int i = 5; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 5; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 5, false, false));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 6, false, false));
@@ -117,7 +117,7 @@ public class cgPPC02 extends TestCgPPC {
 		assertTrue("wrong join", checkJoin(getJoin(5), 0, 34, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(7), 0, 36, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(8), 1, 25, vol, false));
-		for (int i = 9; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 9; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 7, false, false));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 19, false, false));
@@ -136,7 +136,7 @@ public class cgPPC02 extends TestCgPPC {
 		assertNull("wrong join", getJoin(2));
 		assertTrue("wrong join", checkJoin(getJoin(3), 0, 25, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(4), 8, 24, vol, false));
-		for (int i = 5; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 5; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 6, false, false));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 17, false, false));
