+++ junitHostTest/ch/ntb/inf/deep/comp/hosttest/cgPPC/cgPPC03.java
@@ -24,7 +24,7 @@ import static org.junit.Assert.assertTrue;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import ch.ntb.inf.deep.cg.ppc.RegAllocator;
+import ch.ntb.inf.deep.cg.ppc.RegAllocatorPPC;
 import ch.ntb.inf.deep.classItems.CFR;
 import ch.ntb.inf.deep.classItems.Class;
 import ch.ntb.inf.deep.config.Configuration;
@@ -45,7 +45,7 @@ public class cgPPC03 extends TestCgPPC {
 	@Test
 	public void switchNear1() {
 		getCode("switchNear1");
-		for (int i = 0; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 0; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 	
@@ -56,7 +56,7 @@ public class cgPPC03 extends TestCgPPC {
 		assertNull("wrong join", getJoin(0));
 		assertNull("wrong join", getJoin(1));
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 21, vol, false));
-		for (int i = 3; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 3; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 	
@@ -65,7 +65,7 @@ public class cgPPC03 extends TestCgPPC {
 	public void switchNear3() {
 		getCode("switchNear3");
 		assertNull("wrong join", getJoin(0));
-		for (int i = 0; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 0; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 	
@@ -73,7 +73,7 @@ public class cgPPC03 extends TestCgPPC {
 	@Test
 	public void switchFar1() {
 		getCode("switchFar1");
-		for (int i = 0; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 0; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 
@@ -83,7 +83,7 @@ public class cgPPC03 extends TestCgPPC {
 		getCode("switchFar2");
 		assertNull("wrong join", getJoin(0));
 		assertTrue("wrong join", checkJoin(getJoin(1), 0, 20, vol, false));
-		for (int i = 2; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 2; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 
@@ -93,7 +93,7 @@ public class cgPPC03 extends TestCgPPC {
 		assertNull("wrong join", getJoin(0));
 		assertNull("wrong join", getJoin(1));
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 10, vol, false));
-		for (int i = 3; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 3; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
