+++ junitHostTest/ch/ntb/inf/deep/comp/hosttest/cgPPC/cgPPC04.java
@@ -24,7 +24,7 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 
 import ch.ntb.inf.deep.cg.Code32;
-import ch.ntb.inf.deep.cg.ppc.RegAllocator;
+import ch.ntb.inf.deep.cg.ppc.RegAllocatorPPC;
 import ch.ntb.inf.deep.classItems.CFR;
 import ch.ntb.inf.deep.classItems.Class;
 import ch.ntb.inf.deep.config.Configuration;
@@ -48,7 +48,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertNull("wrong join", getJoin(0));
 		assertNull("wrong join", getJoin(1));
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 8, vol, false));
-		for (int i = 3; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 3; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 1, false, false));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 2, true, false));
@@ -64,7 +64,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 24, vol, true));
 		assertTrue("wrong join", checkJoin(getJoin(2).next, 26, 29, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(3), 1, 24, vol, false));
-		for (int i = 4; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 4; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 2, false, false));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 3, false, false));
@@ -82,7 +82,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertNull("wrong join", getJoin(0));
 		assertNull("wrong join", getJoin(1));
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 7, vol, false));
-		for (int i = 3; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 3; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 4, false, false));
 	}
@@ -94,7 +94,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertNull("wrong join", getJoin(0));
 		assertNull("wrong join", getJoin(1));
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 5, vol, false));
-		for (int i = 3; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 3; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 1, true, true));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 2, true, false));
@@ -104,7 +104,7 @@ public class cgPPC04 extends TestCgPPC {
 	@Test
 	public void whileTrueBreak() {
 		getCode("whileTrueBreak");
-		for (int i = 0; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 0; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 
@@ -115,7 +115,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertNull("wrong join", getJoin(0));
 		assertNull("wrong join", getJoin(1));
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 8, vol, false));
-		for (int i = 3; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 3; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 4, false, false));
 	}
@@ -128,7 +128,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertNull("wrong join", getJoin(1));
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 10, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(3), 1, 10, vol, false));
-		for (int i = 4; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 4; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 7, false, false));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 8, false, false));
@@ -143,7 +143,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertNull("wrong join", getJoin(2));
 		assertTrue("wrong join", checkJoin(getJoin(3), 0, 17, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(4), 0, 16, vol, false));
-		for (int i = 5; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 5; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 6, true, false));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 7, false, false));
@@ -167,7 +167,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertNull("wrong join", getJoin(1));
 		assertNull("wrong join", getJoin(2));
 		assertTrue("wrong join", checkJoin(getJoin(3), 0, 19, vol, false));
-		for (int i = 4; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 4; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 	
@@ -221,7 +221,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertTrue("wrong join", checkJoin(getJoin(5), 6, 18, vol, true));	// i
 		assertTrue("wrong join", checkJoin(getJoin(5).next, 19, 64, vol, false));	// i
 												// index 6, valid, keine joins 
-		for (int i = 6; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 6; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 
@@ -229,7 +229,7 @@ public class cgPPC04 extends TestCgPPC {
 	@Test
 	public void phiFunctionTest1() {
 		Code32 code = getCode("phiFunctionTest1");
-		for (int i = 0; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 0; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 0, true, false));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 1, true, false));
@@ -239,7 +239,7 @@ public class cgPPC04 extends TestCgPPC {
 	@Test
 	public void phiFunctionTest2() {
 		Code32 code = getCode("phiFunctionTest2");
-		for (int i = 0; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 0; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 1, true, false));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 2, true, false));
@@ -252,7 +252,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertNull("wrong join", getJoin(0));
 		assertNull("wrong join", getJoin(1));
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 7, vol, false));
-		for (int i = 3; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 3; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 1, false, false));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 2, true, false));
@@ -265,7 +265,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertNull("wrong join", getJoin(0));
 		assertNull("wrong join", getJoin(1));
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 5, vol, false));
-		for (int i = 3; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 3; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 1, true, true));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 2, true, false));
@@ -278,7 +278,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertNull("wrong join", getJoin(0));
 		assertNull("wrong join", getJoin(1));
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 6, vol, false));
-		for (int i = 3; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 3; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 1, false, false));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 2, true, false));
@@ -293,7 +293,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertTrue("wrong join", checkJoin(getJoin(2), 8, 18, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(3), 0, 7, vol, true));
 		assertTrue("wrong join", checkJoin(getJoin(3).next, 9, 16, vol, false));
-		for (int i = 4; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 4; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 4, true, false));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 5, false, false));
@@ -310,7 +310,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 18, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(3), 1, 8, vol, true));
 		assertTrue("wrong join", checkJoin(getJoin(3).next, 9, 16, vol, false));
-		for (int i = 4; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 4; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 5, true, true));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 6, false, false));
@@ -327,7 +327,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 22, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(3), 1, 9, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(4), 12, 20, vol, false));
-		for (int i = 5; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 5; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 5, true, true));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 6, false, false));
@@ -347,7 +347,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertTrue("wrong join", checkJoin(getJoin(3), 0, 23, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(4), 1, 10, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(5), 13, 21, vol, false));
-		for (int i = 6; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 6; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 5, true, true));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 6, false, false));
@@ -368,7 +368,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertTrue("wrong join", checkJoin(getJoin(3), 10, 22, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(4), 1, 9, vol, true));
 		assertTrue("wrong join", checkJoin(getJoin(4).next, 11, 21, vol, false));
-		for (int i = 5; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 5; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 5, true, true));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 6, true, false));
@@ -385,7 +385,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertNull("wrong join", getJoin(0));
 		assertNull("wrong join", getJoin(1));
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 7, vol, false));
-		for (int i = 3; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 3; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 0, true, true));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 1, true, false));
@@ -399,7 +399,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertNull("wrong join", getJoin(1));
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 5, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(3), 0, 5, vol, false));
-		for (int i = 4; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 4; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 2, true, true));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 3, false, false));
@@ -414,7 +414,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertTrue("wrong join", checkJoin(getJoin(2), 11, 23, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(3), 2, 11, vol, true));
 		assertTrue("wrong join", checkJoin(getJoin(3).next, 13, 21, vol, false));
-		for (int i = 4; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 4; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 6, true, false));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 7, false, false));
@@ -445,7 +445,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertTrue("wrong join", checkJoin(getJoin(3), 0, 31, vol, false));	// a
 		assertTrue("wrong join", checkJoin(getJoin(4), 0, 14, vol, true));	// b
 		assertTrue("wrong join", checkJoin(getJoin(4).next, 14, 33, vol, false));
-		for (int i = 5; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 5; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 
@@ -470,7 +470,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertNull("wrong join", getJoin(2));
 		assertTrue("wrong join", checkJoin(getJoin(3), 0, 27, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(4), 12, 29, vol, false));
-		for (int i = 5; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 5; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 
@@ -483,7 +483,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertNull("wrong join", getJoin(2));
 		assertTrue("wrong join", checkJoin(getJoin(3), 0, 26, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(4), 11, 21, vol, false));
-		for (int i = 5; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 5; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 5, true, false));
 		assertTrue("wrong phi function", checkPhiFunction(code.ssa, 6, false, false));
@@ -517,7 +517,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertTrue("wrong join", checkJoin(getJoin(13).next, 64, 98, nonVol, false));	// addr
 		assertTrue("wrong join", checkJoin(getJoin(14), 65, 98, nonVol, false));	// i
 		assertNull("wrong join", getJoin(16));	// elemAddr
-		for (int i = 17; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 17; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 	
@@ -540,7 +540,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertTrue("wrong join", checkJoin(getJoin(11), 1, 101, nonVol, false));	// m
 		assertTrue("wrong join", checkJoin(getJoin(12), 6, 24, nonVol, true));	// i
 		assertTrue("wrong join", checkJoin(getJoin(12).next, 80, 101, vol, false));	// i
-		for (int i = 13; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 13; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 
@@ -568,7 +568,7 @@ public class cgPPC04 extends TestCgPPC {
 		assertTrue("wrong join", checkJoin(getJoin(4).next, 24, 33, vol, false));	// n
 		assertTrue("wrong join", checkJoin(getJoin(5), 4, 15, vol, false));	// b
 		assertTrue("wrong join", checkJoin(getJoin(6), 5, 15, vol, false));	// k
-		for (int i = 7; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 7; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 }
