+++ junitHostTest/ch/ntb/inf/deep/comp/hosttest/cgPPC/cgPPC06.java
@@ -24,7 +24,7 @@ import static org.junit.Assert.assertTrue;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import ch.ntb.inf.deep.cg.ppc.RegAllocator;
+import ch.ntb.inf.deep.cg.ppc.RegAllocatorPPC;
 import ch.ntb.inf.deep.classItems.CFR;
 import ch.ntb.inf.deep.classItems.Class;
 import ch.ntb.inf.deep.config.Configuration;
@@ -48,7 +48,7 @@ public class cgPPC06 extends TestCgPPC {
 		getCode("conditionalOperator1");
 		assertTrue("wrong join", checkJoin(getJoin(0), 5, 9, vol, true));
 		assertTrue("wrong join", checkJoin(getJoin(0).next, 13, 22, vol, false));
-		for (int i = 1; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 1; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 	
@@ -58,7 +58,7 @@ public class cgPPC06 extends TestCgPPC {
 		getCode("conditionalOperator2");
 		assertTrue("wrong join", checkJoin(getJoin(0), 9, 13, vol, true));
 		assertTrue("wrong join", checkJoin(getJoin(0).next, 17, 26, vol, false));
-		for (int i = 1; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 1; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 	
@@ -67,7 +67,7 @@ public class cgPPC06 extends TestCgPPC {
 	public void conditionalOperator3() {
 		getCode("conditionalOperator3");
 		assertTrue("wrong join", checkJoin(getJoin(0), 5, 9, vol, false));
-		for (int i = 1; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 1; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
 
@@ -76,7 +76,7 @@ public class cgPPC06 extends TestCgPPC {
 	public void conditionalOperator4() {
 		getCode("conditionalOperator4");
 		assertTrue("wrong join", checkJoin(getJoin(0), 4, 8, vol, false));
-		for (int i = 1; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 1; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
