+++ junitHostTest/ch/ntb/inf/deep/comp/hosttest/cgPPC/cgPPC08.java
@@ -24,7 +24,7 @@ import static org.junit.Assert.assertTrue;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import ch.ntb.inf.deep.cg.ppc.RegAllocator;
+import ch.ntb.inf.deep.cg.ppc.RegAllocatorPPC;
 import ch.ntb.inf.deep.classItems.CFR;
 import ch.ntb.inf.deep.classItems.Class;
 import ch.ntb.inf.deep.config.Configuration;
@@ -50,7 +50,7 @@ public class cgPPC08 extends TestCgPPC {
 		assertTrue("wrong join", checkJoin(getJoin(2), 0, 11, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(3), 0, 12, vol, false));
 		assertTrue("wrong join", checkJoin(getJoin(4), 1, 11, vol, false));
-		for (int i = 5; i < RegAllocator.maxNofJoins; i++)
+		for (int i = 5; i < RegAllocatorPPC.maxNofJoins; i++)
 			assertNull("wrong join", getJoin(i));
 	}
