+++ src/ch/ntb/inf/deep/cg/ppc/CodeGenPPC.java
@@ -33,7 +33,6 @@ import ch.ntb.inf.deep.strings.HString;
 
 public class CodeGenPPC extends CodeGen implements InstructionOpcs, Registers {
 
-	static final int maxNofParam = 32;
 	private static final int arrayLenOffset = 6;	
 	private static final int tempStorageSize = 48;	// 1 FPR(temp), 4 GPRs 
 	
@@ -68,8 +67,6 @@ public class CodeGenPPC extends CodeGen implements InstructionOpcs, Registers {
 	static boolean[] paramHasNonVolReg = new boolean[maxNofParam];
 	// register of parameter, long and doubles count as 2 parameters
 	static int[] paramRegNr = new int[maxNofParam];
-	// last instruction where parameters is used
-	static int[] paramRegEnd = new int[maxNofParam];
 	
 	// information about into which registers parameters of this method go 
 	private static int nofMoveGPR, nofMoveFPR;
@@ -123,7 +120,7 @@ public class CodeGenPPC extends CodeGen implements InstructionOpcs, Registers {
 		if (dbg) StdStreams.vrb.println("build intervals");
 //		StdStreams.vrb.print(ssa.cfg.toString());
 //		StdStreams.vrb.println(ssa.toString());
-		RegAllocator.buildIntervals(ssa);
+		RegAllocatorPPC.buildIntervals(ssa);
 		
 		if (dbg) StdStreams.vrb.println("assign registers to parameters");
 		SSANode b = (SSANode) ssa.cfg.rootNode;
@@ -140,10 +137,10 @@ public class CodeGenPPC extends CodeGen implements InstructionOpcs, Registers {
 		}
 		
 		if (dbg) StdStreams.vrb.println("allocate registers");
-		RegAllocator.assignRegisters();
+		RegAllocatorPPC.assignRegisters();
 		
 		if (dbg) {
-			StdStreams.vrb.println(RegAllocator.joinsToString());
+			StdStreams.vrb.println(RegAllocatorPPC.joinsToString());
 		}
 		if (dbg) {
 			StdStreams.vrb.print("register usage in method: nofNonVolGPR = " + nofNonVolGPR + ", nofVolGPR = " + nofVolGPR);
@@ -247,8 +244,8 @@ public class CodeGenPPC extends CodeGen implements InstructionOpcs, Registers {
 				if (exitSet[i+maxStackSlots] != null) {	// if null -> parameter is never used
 					if(dbg) StdStreams.vrb.print("r");
 					if (paramHasNonVolReg[i]) {
-						int reg = RegAllocator.reserveReg(gpr, true);
-						int regLong = RegAllocator.reserveReg(gpr, true);
+						int reg = RegAllocatorPPC.reserveReg(gpr, true);
+						int regLong = RegAllocatorPPC.reserveReg(gpr, true);
 						moveGPRsrc[nofMoveGPR] = nofParamGPR;
 						moveGPRsrc[nofMoveGPR+1] = nofParamGPR+1;
 						moveGPRdst[nofMoveGPR++] = reg;
@@ -258,16 +255,16 @@ public class CodeGenPPC extends CodeGen implements InstructionOpcs, Registers {
 						if(dbg) StdStreams.vrb.print(reg + ",r" + regLong);
 					} else {
 						int reg = paramStartGPR + nofParamGPR;
-						if (reg <= paramEndGPR) RegAllocator.reserveReg(gpr, reg);
+						if (reg <= paramEndGPR) RegAllocatorPPC.reserveReg(gpr, reg);
 						else {
-							reg = RegAllocator.reserveReg(gpr, false);
+							reg = RegAllocatorPPC.reserveReg(gpr, false);
 							moveGPRsrc[nofMoveGPR] = nofParamGPR;
 							moveGPRdst[nofMoveGPR++] = reg;
 						}
 						int regLong = paramStartGPR + nofParamGPR + 1;
-						if (regLong <= paramEndGPR) RegAllocator.reserveReg(gpr, regLong);
+						if (regLong <= paramEndGPR) RegAllocatorPPC.reserveReg(gpr, regLong);
 						else {
-							regLong = RegAllocator.reserveReg(gpr, false);
+							regLong = RegAllocatorPPC.reserveReg(gpr, false);
 							moveGPRsrc[nofMoveGPR] = nofParamGPR + 1;
 							moveGPRdst[nofMoveGPR++] = regLong;
 						}
@@ -282,16 +279,16 @@ public class CodeGenPPC extends CodeGen implements InstructionOpcs, Registers {
 				if (exitSet[i+maxStackSlots] != null) {	// if null -> parameter is never used
 					if(dbg) StdStreams.vrb.print("fr");
 					if (paramHasNonVolReg[i]) {
-						int reg = RegAllocator.reserveReg(fpr, true);
+						int reg = RegAllocatorPPC.reserveReg(fpr, true);
 						moveFPRsrc[nofMoveFPR] = nofParamFPR;
 						moveFPRdst[nofMoveFPR++] = reg;
 						paramRegNr[i] = reg;
 						if(dbg) StdStreams.vrb.print(reg);
 					} else {
 						int reg = paramStartFPR + nofParamFPR;
-						if (reg <= paramEndFPR) RegAllocator.reserveReg(fpr, reg);
+						if (reg <= paramEndFPR) RegAllocatorPPC.reserveReg(fpr, reg);
 						else {
-							reg = RegAllocator.reserveReg(fpr, false);
+							reg = RegAllocatorPPC.reserveReg(fpr, false);
 							moveFPRsrc[nofMoveFPR] = nofParamFPR;
 							moveFPRdst[nofMoveFPR++] = reg;
 						}
@@ -305,16 +302,16 @@ public class CodeGenPPC extends CodeGen implements InstructionOpcs, Registers {
 				if (exitSet[i+maxStackSlots] != null) {	// if null -> parameter is never used
 					if(dbg) StdStreams.vrb.print("r");
 					if (paramHasNonVolReg[i]) {
-						int reg = RegAllocator.reserveReg(gpr, true);
+						int reg = RegAllocatorPPC.reserveReg(gpr, true);
 						moveGPRsrc[nofMoveGPR] = nofParamGPR;
 						moveGPRdst[nofMoveGPR++] = reg;
 						paramRegNr[i] = reg;
 						if(dbg) StdStreams.vrb.print(reg);
 					} else {
 						int reg = paramStartGPR + nofParamGPR;
-						if (reg <= paramEndGPR) RegAllocator.reserveReg(gpr, reg);
+						if (reg <= paramEndGPR) RegAllocatorPPC.reserveReg(gpr, reg);
 						else {
-							reg = RegAllocator.reserveReg(gpr, false);
+							reg = RegAllocatorPPC.reserveReg(gpr, false);
 							moveGPRsrc[nofMoveGPR] = nofParamGPR;
 							moveGPRdst[nofMoveGPR++] = reg;
 						}
